// src/generator/dmmf/types.ts
export type ReadonlyDeep<O> = {
  +readonly [K in keyof O]: ReadonlyDeep<O[K]>;
};

export namespace DMMF {
// ...
  export type SchemaArg = ReadonlyDeep<{
    name: string;
    comment?: string;
    isNullable: boolean;
    isRequired: boolean;
    // inputTypes: InputTypeRef[];
    deprecation?: Deprecation;
    // additional props:
    selectedInputType: SchemaArgInputType;
    typeName: string;
    typeGraphQLType: string;
    fieldTSType: string;
    hasMappedName: boolean;
    isOmitted: boolean;
    isOptional: boolean;
  }>;
  export type OutputType = ReadonlyDeep<{
    name: string;
    // fields: SchemaField[];
    // additional props:
    fields: OutputSchemaField[];
    typeName: string;
  }>;
// ...
  export type OutputSchemaField = SchemaField &
    ReadonlyDeep<{
      argsTypeName: string | undefined;
    }>;
  export type InputType = ReadonlyDeep<{
    name: string;
    constraints: {
      maxNumFields: number | null;
      minNumFields: number | null;
    };
    meta?: {
      source?: string;
    };
    fields: SchemaArg[];
    // additional props:
    typeName: string;
  }>;
// ...
}